requirements.txt
streamlit
llama-index
PyPDF2
python-dotenv